FunkyRhythm Song, New Dig Move, Arrow Items, and Moblin Move Logic Changes
I added a new song to our audio collection. I also added arrow items
that the player must have in their inventory to use the bow attack. The
Moblin now have a line of sight that they can use to find the player if
within a distance. If they see the player, then they will chase them if
on ‘Hard Mode’. Lastly, I added a dig attack that the player can use to
remove rocks from the game board. There is a probability of receiving
loot when completing this action.
30-69.9999% => Small Rupee
70-94.9999% => Medium Rupee
95-100% => Large Rupee