In Game Shop With Merchant, Shop Canvas and Panel With Options, Merchant Animations, and ShopScript for Shop Logic
I added an in-game shop that the player can purchase items from. The
merchant is located around the top lining of the game board (right
around the middle of the top row). I also added logic for this shop in
a new script called ‘ShopScript’. The merchant sells arrows, bombs, and
experience. The merchant does an animation when the player leaves the
shop screen.