Handled Cases Where Player Can Move In Shop and Pause Screens
I worked on handling a couple of bugs where the player could move while
in the shop or when the game is paused. I also worked on toggling the
display of the special attack menu. There still is a small case where
it doesn’t function accordingly, but it doesn’t affect the gameplay at
all. Specifically, the case is where the player is in the shop and
opens and closes the pause screen. The special attack menu should be
hidden, but is displayed to the user.